CLI: overhaul repro command
- sb repro as it was is entirely removed
- sb repro-next becomes sb repro
