+++ code/package.json
@@ -30,8 +30,7 @@
     "ci-tests": "yarn task --task check --no-link --start-from=install && yarn lint && yarn test && cd ../scripts && yarn test",
     "coverage": "codecov",
     "danger": "danger",
-    "generate-repros": "zx ../scripts/repros-generator/index.mjs",
-    "generate-repros-next": "ts-node --swc ../scripts/next-repro-generators/generate-repros.ts",
+    "generate-repros": "ts-node --swc ../scripts/next-repro-generators/generate-repros.ts",
     "github-release": "github-release-from-changelog",
     "linear-export": "ts-node --swc --project=../scripts/tsconfig.json ../scripts/linear-export.ts",
     "lint": "yarn lint:js && yarn lint:md",
